Add UseDigestTo* marker trait and blanket Sign/Verify impls
This commit makes the digest for `SignDigest` and `VerifyDigest` an
associated type rather than a generic, and also adds two marker traits:

- `UseDigestToSign`: opt into blanket impl of `Sign` for `SignDigest`
- `UseDigestToVerify`: opt into blanket impl of `Verify` for `VerifyDigest`

The reason for requiring a marker trait is there are cases where the
same keys can be used with an IUF mode of an algorithm which is distinct
from the non-IUF form, e.g. Ed25519 vs Ed25519ph. In this case, the
underlying keys are the same, but the signature algorithm when using the
IUF mode is distinct.