Add "raw digest" signature marker trait and sign/verify traits
Introduces the notion of a "raw digest" signature algorithm, i.e.
any algorithm where signatures are always computed as `S(H(m)))` where:

- `S`: signature algorithm
- `H`: hash (a.k.a. digest) function
- `m`: message

Notably this does not hold true for Ed25519, which hashes the input
message twice in an effort to remain secure even in the event of
collisions in the underlying hash function.

However, it supports a separate IUF mode Ed25519ph, which is trivial
for any Ed25519 implementation to implement (it hashes the message
in advance, then performs a regular Ed25519 signature albeit with a
domain separation tweak).

For cases like this, where the IUF mode is deliberately domain separated
from the normal mode, it would be nice to avoid a blanket impl which
assumes all signature algorithms are composed as `S(H(m))`.

To accomplish this, this commit adds a `RawDigestSignature` marker trait
to be applied at the algorithm level to the corresponding signature
trait.

For signature types marked as `RawDigestSignature`, it's possible to
`impl` a corresponding `SignRawDigest` trait for which a blanket `impl`
of `Sign` exists which hashes the message with the corresponding
`Digest` algorithm.

Additionally, `SignDigest` is blanket impl'd for all `SignRawDigest`
types.