Add `try_*` methods to `*Signer` traits
For many signature providers, it's not possible for an error to occur
when creating a signature. Support for handling errors which occur when
computing signatures is intended for use with cloud KMS, HSMs, or other
hardware tokens where things like I/O errors are possible.

To simplify usage in the event that only software signers are in use,
this commit splits the `sign` API into `Result`-based `try_sign` method
and another `sign` method that always assumes success and unwraps the
result of the former method.