signature: Remove `alloc` feature; MSRV 1.31+
The only thing the `signature` crate uses `alloc` for is the
`Signature::to_vec` method, which is trivially accomplished otherwise as
`sig.as_ref().to_vec()` or various other ways.

By getting rid of it, we can completely get rid of the `alloc` feature,
and with that reduce the MSRV back to 1.31.

It might be worth considering a 1.0 release with a higher MSRV so we can
leverage `TryFrom`, but for now, this provides wider compatibility by
removing a single (mis)feature.