ed25519: Perform partial reduction check on `Signature`
Ensures that the three highest bits of the `s` scalar component of an
Ed25519 signature are unset.

This doesn't ensure that `s` is fully reduced (which would require a
full reduction check in the event that the 4th most significant bit is
set), however it will catch a number of invalid signatures relatively
cheaply.

Inspired by:

https://github.com/dalek-cryptography/ed25519-dalek/pull/99