elliptic-curve-crate: Extract from `ecdsa` crate
The `ecdsa` crate contains types, traits, and a complete registry for
common Weierstrass curves which is reusable for both the signing and
encryption use case.

This extracts the reusable functionality into the `elliptic-curve`
crate, with the goal of rewriting the duplicated parts of the current
`ecdsa` crate to use types from the `elliptic-curve` crate, and once
that's done, finding a better home, e.g.
https://github.com/RustCrypto/traits

It seems we collectively own the following crates:

- `curve25519`
- `ed25519`
- `p256`
- `p384`
- `p521`
- `k256`

...so it would be nice to extract the curve registry in these crates
into those crates repsectively, e.g. under a hypothetical
https://github.com/RustCrypto/elliptic-curves

But for now, this structures the crate the same way the `ecdsa` crate
was previously structured: with an all-in-one curve registry contained
within the crate.