ed25519: avoid serializing a length prefix with serde
Previously the serde serializer used `Serialize::serialize_bytes` which
accepts a slice and when serialized with e.g. bincode also includes a
length prefix on the signature. As it were, this made it incompatible
with the extant deserializer (and as such, this PR also adds much needed
serialization tests).

This commit changes the serializer to use the same internal
implementation as serde's own `array_impls`, i.e. signatures are
serialized as a tuple with 64 members.

Using this implementation, it means an Ed25519 signature serialized with
bincode is identical to `to_bytes()`, i.e. there is no redundant length
prefix.