Cargo.toml: fetch RustCrypto/elliptic-curves crates from git
Adds a `[patch.crates-io]` directive so we can work off the latest
versions of the https://github.com/RustCrypto/elliptic-curves crates.

This is a breaking change for the `ecdsa` crate and makes it
unreleasable until releases of the elliptic curve crates happen,
but will allow us to implement a number of new features which require a
curve arithmetic implementation.