ecdsa: bump version to v0.7.0-pre
This commit isn't intended to be a release, but just bumps the version
number in Cargo.toml to `-pre` to distinguish it from the currently
released v0.6.1.

Several backwards incompatible / semver breaking changes have been
introduced since the release.

It also updates `elliptic-curve` to v0.5.0-pre, which was similarly
bumped to a `-pre` version in:

https://github.com/RustCrypto/traits/pull/215