ecdsa: add DigestPrimitive
The `DigestPrimitive` trait, intended to be implemented within a crate
which impls `elliptic_curve::weierstrass::Curve`, provides a blanket
impl of `PrehashSignature` for `ecdsa::Signature<C: Curve>`.

This is a small hack needed to work around the fact that the ECDSA
signature trait now lies in an external crate, so crates which implement
a specific curve can no longer impl traits on `ecdsa::Signature`.