ecdsa: initial `dev` module with `new_*_test` macros
Adds a `dev` module ala the ones in the trait crates which contains
`new_signing_test` and `new_verification_test` macros which implement
a basic set of tests for implementers of `SignPrimitive` and
`VerifyPrimitive` respectively.

Ideally this would implement something close to the set of tests
described in the FIPS 186-4 ECDSA Validation System (ECDSA2VS), however
for now it tests basic signing, verification, and a verification failure
in the event the `s` component of the signature has an incorrect value.

Eventually it'd probably be good to convert to using `blobby` for
consistency with the other `dev` modules, however for now it specifies a
`TestVector` struct which the macros accept a slice of for vectors.