ecdsa: add CheckSignatureBytes trait + NonZeroScalar components
Adds a `CheckSignatureBytes` trait with a default impl that ensures
neither the `r` nor `s` components of a signature are all-zero.

Includes a blanket impl of `CheckSignatureBytes` for all curves with
the `Arithmetic` trait defined (and thus access to `C::Scalar`). This
performs a better check that both `r` and `s` do not overflow the
curve's order and are non-zero (using `NonZeroScalar`).

Together these allow conditionally definining `ecdsa::Signature::r` and
`::s` methods which permit infallible access to `NonZeroScalar<C>`,
rather than the previous implementation which returns bytes.

All of the above is a compromise which allows for crates which do not
provide a curve arithmetic backend to still assert some checks on the
signature values, but also ensure that signatures are well-formed,
in-range scalars "automatically" when arithmetic is available.