ed25519: add (doctested!) `ed25519-dalek` usage example
Adds a rustdoc example which completely demonstrates how to use code
written generically with `ed25519::Signature` and `signature::{Signer,
Verifier}` with the concrete `ed25519_dalek::{Keypair, PublicKey}`
types.

Unlike the previous example, which hardly showed anything, this commit
also pulls in `ed25519-dalek` as a `[dev-dependency]`, which makes it
possible to actually doctest the example.