feat(ed25519): Add a feature to serialize with serde's bytes type
This introduces a new feature, `serde_as_bytes`, that switches the
implementation of (de)serialization to use serde's `*_bytes` methods,
allowing crate consumers to opt-in to this approach which is beneficial
for formats with optimised byte array representations.

This has required a small change to the test workflow, since it's
necessary to test both with and without the `serde_as_bytes` feature.
The features cannot be made mutually exclusive, though logically they
are, without changing the `serde` feature flag, which would be
backwards-incompatible.

Fixes #336.