ecdsa: add `RecoveryId` type (#392)
Previously `RecoverableSignPrimitive` would return a 1-bit flag to use
as the "recid".

While this suffices for the only current use case of this crate, which
is secp256k1, it doesn't for other curves, where there is a practical
possibility that an affine coordinate (i.e. element of the base field)
might overflow the curve's order, i.e. that it won't be a valid element
of the scalar field.

This commit adds a 0-3 valued `RecoveryId` type which captures all
cases, and uses it as the return value for `RecoverableSignPrimitive`.
This provides a cross-curve compatible type for the recovery ID,
which can be used in a generic implementation of ECDSA (forthcoming).