ecdsa: add `MulBaseReduced` trait
Adds a helper trait for performing the 𝑘×𝑮 scalar multiplication used by
ECDSA, with a subsequent reduction to a scalar field element.

The main purpose of having a special trait for this is so curve
implementations can avoid exposing their base field implementations.
This trait does just what's needed with base field elements for ECDSA.

With the addition of this trait, it's possible to have a generic
implementation of ECDSA signatures which otherwise only relies on the
existing curve arithmetic features.