Add default impl of `SignPrimitive::try_sign_prehashed`
Leverages the newly added `AffineXCoordinate` trait from
RustCrypto/traits#817 to provide a default generic implementation of
the ECDSA signature algorithm.

Downstream crates can choose to override this implementation if they so
desire, allowing them to potentially leverage more efficient arithmetic.

This approach does have a disadvantage that it necessarily adds a
`ScalarArithmetic<Scalar = Self>` bound on the curve, meaning that
`SignPrimitive` can only be impl'd for a particular curve's scalar type,
as opposed to the original design goal of allowing it to be impl'd for a
hardware ECDSA accelerator that stores a private scalar.

Either the default impl needs to be replaced with a blanket impl with
more restrictive bounds, or a separate trait needs to be added to
support hardware accelerators. The latter may indeed make the most
sense.