ed25519: deprecate `Signature::new` in favor of `::from_bytes` (#401)
Adds a `Signature::from_bytes` inherent method which matches the method
of the same name on the `Signature` trait, i.e. it returns a
`signature::Result`.

This method performs a partial range check on the `s` component of the
signature to ensure fit within the order of ~2^(252.5). It won't reject
all invalid signatures, but it will detect quite a few of them.

Deprecates the infallible `Signature::new` method, since it can't detect
invalid signatures. While the `from_bytes` method isn't perfect, it can
be improved in the future, whereas an infallible method can't.