ecdsa: impl `signature::hazmat::{PrehashSigner, PrehashVerifier}` (#534)
Adds "hazmat" trait impls to `SigningKey` and `VerifyingKey`
respectively which allow computing signatures over raw message digests.
See RustCrypto/traits#1099.

The implementation allows digests which are shorter or longer than the
field size of the curve, using zero-padding if the digest is too short,
and truncating if it's too long. The minimum digest size is set to half
of the curve's field size.