ecdsa: refactor RFC6979 usage and trait impls
Uses `C::Digest` unilaterally as the digest function to instantiate
HMAC-DRBG for RFC6979. This replaces the previous inconsistent usage
where sometimes `C::Digest` was used while in `DigestSigner` the generic
`D` parameter was used.

With this, the various trait impls can be composed one in terms of the
other:

    `*Signer` -> `*DigestSigner` -> `*PrehashSigner`

This commit also removes `SignPrimitive::try_sign_digest_rfc6979` as the
`*PrehashSigner` impls are now the only caller into `SignPrimitive` and
the `DigestSigner` impls handle hashing.