ecdsa: add RFC5758 OID support
Adds support for obtaining the RFC5758 OIDs associated with the digest
algorithm used with the `Signer` and `Verifier` traits, i.e. the digest
defined via `DigestPrimitive::Digest` for a given curve.

The ECDSA OID is available via an `AssociatedOid` impl on `Signature`.

The `AlgorithmIdentifier` is also available via the newly added
`AssociatedAlgorithmIdentifier` trait.