ecdsa: fix `dev` macros
The method changed from `GenericArray::from_slice` to
`Array::ref_from_slice`.

This has been renamed back upstream in `hybrid-array` to simplify
people's upgrades, but for now we need to rename it until we can upgrade
`hybrid-array` again.

Longer term the method will be deprecated, so this is temporary anyway.