dsa: add `from_components_unchecked` APIs (#1059)
Adds support for any size of keys, either larger or older/weak keys.

The API is only available under the `hazmat` feature-flag, and the
consumer is responsible for checking the security of what they're using
it for.

Fixes #1058

* dsa: compare key sizes aligned to limb

This allows to compare key sizes after a back and forth through
serialization which could lose the exact key size.